// src/modules/districts/dtos/districts-dto.ts
export type PenaltyData = {
  uniform: number;
  timeOver: number;
  misconduct: number;
  ieadcConsuetude: number;
  musiciansIsNotSixtyPercentOfTeens: number;
  totalPenalty: number;
};

export type DesqualificationData = {
  membersAreInvaders: boolean;
  noRecommendationLetter: boolean;
  singersAreAnotherRegional: boolean;
  musiciansAreIsAnotherRegional: boolean;
  moreThanTwoMusiciansFromUniaadcBase: boolean;
  membersOfGeneralBackingWillBeAblePerformSolosOrComposeRegionalBacking: boolean;
};